django/requirements.txt
Django==3.2
djangorestframework
mysqlclient
django-mysql
django-cors-headers
python-dotenv